requirements.txt
pandas
numpy
opencv-python
torch
torchvision
scipy
Pillow
PyYAML